api: move API related data structure into nydus-api crate
Move API related data structure into nydus-api crate.
