+++ api/Cargo.toml
@@ -9,6 +9,7 @@ license = "Apache-2.0 OR BSD-3-Clause"
 
 [dependencies]
 lazy_static = "1.4.0"
+libc = "0.2"
 log = "0.4.8"
 dbs-uhttp = { version = "0.3.0" }
 serde = { version = "1.0.110", features = ["rc"] }
@@ -17,5 +18,7 @@ serde_json = "1.0.53"
 vmm-sys-util = "0.9.0"
 url = "2.1.1"
 http = "0.2.1"
-nydus-utils = { path = "../utils" }
 mio = { version = "0.8", features = ["os-poll", "os-ext"]}
+
+nydus-error = { version = "0.2.0", path = "../error" }
+nydus-utils = { version = "0.2.0", path = "../utils" }
