+++ rafs/Cargo.toml
@@ -33,6 +33,7 @@ url = { version = "2.1.1", optional = true }
 vm-memory = "0.7.0"
 fuse-backend-rs = { version = "0.9.0" }
 
+nydus-api = { version = "0.1.0", path = "../api" }
 nydus-error = { version = "0.2.0", path = "../error" }
 nydus-storage = { version = "0.5.0", path = "../storage", features = ["backend-localfs"] }
 nydus-utils = { version = "0.2.0", path = "../utils" }
