+++ src/bin/nydusd/blob_cache.rs
@@ -10,10 +10,9 @@ use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::{Arc, Mutex, MutexGuard};
 
 use nydus_api::http::{
-    BlobCacheEntry, BlobCacheList, BlobCacheObjectId, BLOB_CACHE_TYPE_BOOTSTRAP,
+    BlobCacheEntry, BlobCacheList, BlobCacheObjectId, FsCacheConfig, BLOB_CACHE_TYPE_BOOTSTRAP,
 };
 use rafs::metadata::{RafsMode, RafsSuper};
-use storage::cache::FsCacheConfig;
 use storage::device::BlobInfo;
 use storage::factory::{BackendConfig, CacheConfig, FactoryConfig};
