+++ storage/src/backend/mod.rs
@@ -57,65 +57,6 @@ pub enum BackendError {
 /// Specialized `Result` for storage backends.
 pub type BackendResult<T> = std::result::Result<T, BackendError>;
 
-/// Configuration information for network proxy.
-///
-/// This structure is externally visible through configuration file and HTTP API, please keep them
-/// stable.
-#[derive(Debug, Clone, Deserialize, Serialize)]
-#[serde(default)]
-pub struct ProxyConfig {
-    /// Access remote storage backend via P2P proxy, e.g. Dragonfly dfdaemon server URL.
-    url: String,
-    /// Endpoint of P2P proxy health checking.
-    ping_url: String,
-    /// Fallback to remote storage backend if P2P proxy ping failed.
-    fallback: bool,
-    /// Interval of P2P proxy health checking, in seconds.
-    check_interval: u64,
-}
-
-impl Default for ProxyConfig {
-    fn default() -> Self {
-        Self {
-            url: String::new(),
-            ping_url: String::new(),
-            fallback: true,
-            check_interval: 5,
-        }
-    }
-}
-
-/// Generic configuration for storage backends.
-///
-/// This structure is externally visible through configuration file and HTTP API, please keep them
-/// stable.
-#[derive(Debug, Clone, Deserialize, Serialize)]
-#[serde(default)]
-pub struct CommonConfig {
-    /// Enable HTTP proxy for the read request.
-    proxy: ProxyConfig,
-    /// Skip SSL certificate validation for HTTPS scheme.
-    skip_verify: bool,
-    /// Drop the read request once http request timeout, in seconds.
-    timeout: u64,
-    /// Drop the read request once http connection timeout, in seconds.
-    connect_timeout: u64,
-    /// Retry count when read request failed.
-    retry_limit: u8,
-}
-
-impl Default for CommonConfig {
-    fn default() -> Self {
-        Self {
-            proxy: ProxyConfig::default(),
-            skip_verify: false,
-            timeout: 5,
-            connect_timeout: 5,
-            retry_limit: 0,
-        }
-    }
-}
-
 /// Trait to read data from a on storage backend.
 pub trait BlobReader: Send + Sync {
     /// Get size of the blob file.
@@ -242,6 +183,7 @@ fn default_http_scheme() -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use nydus_api::http::RegistryOssConfig;
 
     #[cfg(any(feature = "backend-oss", feature = "backend-registry"))]
     #[test]
@@ -251,7 +193,7 @@ mod tests {
 
     #[test]
     fn test_common_config() {
-        let config = CommonConfig::default();
+        let config = RegistryOssConfig::default();
 
         assert_eq!(config.timeout, 5);
         assert_eq!(config.connect_timeout, 5);
